refactor : Point, StoneColor 패키지 이동
