+++ src/main/kotlin/omok/model/OmokStone.kt
@@ -1,6 +1,3 @@
 package omok.model
 
-import omok.Point
-import omok.StoneColor
-
 data class OmokStone(val point: Point, val color: StoneColor)
