+++ src/main/kotlin/omok/model/StoneColor.kt
@@ -1,4 +1,4 @@
-package omok
+package omok.model
 
 enum class StoneColor {
     BLACK,
