+++ src/test/kotlin/omok/StoneColorTest.kt
@@ -1,5 +1,6 @@
 package omok
 
+import omok.model.StoneColor
 import org.assertj.core.api.AssertionsForClassTypes.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
