+++ src/test/kotlin/omok/fixtures/OmokFixtures.kt
@@ -1,6 +1,6 @@
 package omok.fixtures
 
-import omok.StoneColor
+import omok.model.StoneColor
 import omok.model.OmokStone
 
 fun createOmokStone(
